Add RestApiTeam class
We can't add/remove members to/from teams via the
GraphQL API, so I've added this class.

I'm using net/http rather than something like
Octokit, because I don't want to have to install
gems when we run this as a github action - it
seems like overkill when we only need to make a
single PUT request.